+++ packages/webapp/src/map/queryParams.ts
@@ -1,7 +1,7 @@
 import { useEffect } from 'react'
 import { useMap } from 'react-leaflet'
 
-import { useHistory, useLocation } from 'react-router-dom'
+import { useNavigate, useLocation } from 'react-router-dom'
 
 const throttle = (fn, delay): [() => void, () => void] => {
   let id
@@ -24,14 +24,14 @@ const throttle = (fn, delay): [() => void, () => void] => {
 export const useQueryParams = () => {
   const map = useMap()
   const location = useLocation()
-  const history = useHistory()
+  const navigate = useNavigate()
 
   useEffect(() => {
     const [onChange, onChangeCancel] = throttle(() => {
       const {lat, lng} = map.getCenter()
       const zoom = map.getZoom()
       const search = new URLSearchParams({ lat: lat.toFixed(4), lng: lng.toFixed(4), zoom: `${zoom}` });
-      history.replace({ pathname: location.pathname, search: search.toString() });  
+      navigate(`${location.pathname}?${search.toString()}`, {replace: true});
     }, 200)
 
     map.on('zoomend', onChange)